fc-service-server merge

+++ fc-service-server/src/main/java/eu/gaiax/difs/fc/server/service/ParticipantsService.java
@@ -1,25 +1,31 @@
 package eu.gaiax.difs.fc.server.service;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import eu.gaiax.difs.fc.api.generated.model.Participant;
+import eu.gaiax.difs.fc.api.generated.model.Participants;
+import eu.gaiax.difs.fc.api.generated.model.UserProfile;
+import eu.gaiax.difs.fc.api.generated.model.UserProfiles;
+import eu.gaiax.difs.fc.core.dao.ParticipantDao;
+import eu.gaiax.difs.fc.core.exception.NotFoundException;
+import eu.gaiax.difs.fc.core.pojo.ContentAccessorDirect;
+import eu.gaiax.difs.fc.core.pojo.ParticipantMetaData;
+import eu.gaiax.difs.fc.core.pojo.SelfDescriptionMetadata;
+import eu.gaiax.difs.fc.core.pojo.VerificationResultParticipant;
+import eu.gaiax.difs.fc.core.service.sdstore.SelfDescriptionStore;
+import eu.gaiax.difs.fc.core.service.verification.VerificationService;
+import eu.gaiax.difs.fc.server.generated.controller.ParticipantsApiDelegate;
+import eu.gaiax.difs.fc.server.util.SessionUtils;
 import java.net.URI;
 import java.util.List;
 import java.util.Map;
-
+import java.util.stream.Collectors;
+import javax.ws.rs.ForbiddenException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import eu.gaiax.difs.fc.api.generated.model.Participant;
-import eu.gaiax.difs.fc.api.generated.model.UserProfile;
-import eu.gaiax.difs.fc.core.dao.ParticipantDao;
-import eu.gaiax.difs.fc.core.exception.ClientException;
-import eu.gaiax.difs.fc.core.exception.NotFoundException;
-import eu.gaiax.difs.fc.server.generated.controller.ParticipantsApiDelegate;
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * Implementation of the {@link ParticipantsApiDelegate} interface.
@@ -33,7 +39,15 @@ public class ParticipantsService implements ParticipantsApiDelegate {
   private ParticipantDao partDao;
 
   @Autowired
+  private SelfDescriptionStore selfDescriptionStore;
+
+  @Autowired
+  private VerificationService verificationService;
+
+  @Autowired
   private ObjectMapper jsonMapper;
+
+  private final String catalogueAdminRole="ROLE_Ro-MU-CA";
 
   /**
    * POST /participants : Register a new participant in the catalogue.
@@ -49,8 +63,8 @@ public class ParticipantsService implements ParticipantsApiDelegate {
   @Transactional
   public ResponseEntity<Participant> addParticipant(String body) {
     log.debug("addParticipant.enter; got participant: {}", body); // it can be JWT?
-    Participant part = toParticipant(body);
-    partDao.create(part);
+    ParticipantMetaData part = getParticipantExtWithValidationAndStore(body);
+    part = partDao.create(part);
     log.debug("addParticipant.exit; returning: {}", part);
     return ResponseEntity.created(URI.create("/participants/" + part.getId())).body(part);
   }
@@ -70,7 +84,13 @@ public class ParticipantsService implements ParticipantsApiDelegate {
   @Transactional
   public ResponseEntity<Participant> deleteParticipant(String participantId) {
     log.debug("deleteParticipant.enter; got participant: {}", participantId);
-    Participant part = partDao.delete(participantId)
+    ParticipantMetaData part = partDao.select(participantId)
+        .orElseThrow(() -> new NotFoundException("Participant not found: " + participantId));
+    if(!SessionUtils.sessionUserHasRole(catalogueAdminRole) && !part.getId().equals(SessionUtils.getSessionParticipantId())) {
+      throw new ForbiddenException("User has no permissions to delete participant : " + participantId);
+    }
+    selfDescriptionStore.deleteSelfDescription(part.getSdHash());
+    partDao.delete(participantId)
         .orElseThrow(() -> new NotFoundException("Participant not found: " + participantId));
     log.debug("deleteParticipant.exit; returning: {}", part);
     return ResponseEntity.ok(part);
@@ -90,8 +110,10 @@ public class ParticipantsService implements ParticipantsApiDelegate {
   @Override
   public ResponseEntity<Participant> getParticipant(String participantId) {
     log.debug("getParticipant.enter; got participant: {}", participantId);
-    Participant part = partDao.select(participantId)
+    ParticipantMetaData part = partDao.select(participantId)
         .orElseThrow(() -> new NotFoundException("Participant not found: " + participantId));
+    SelfDescriptionMetadata selfDescriptionMetadata = selfDescriptionStore.getByHash(part.getSdHash());
+    part.setSelfDescription(selfDescriptionMetadata.getSelfDescription().getContentAsString());
     log.debug("getParticipant.exit; returning: {}", part);
     return ResponseEntity.ok(part);
   }
@@ -108,12 +130,13 @@ public class ParticipantsService implements ParticipantsApiDelegate {
    *         Must not outline any information about the internal structure of the server. (status code 500)
    */
   @Override
-  public ResponseEntity<List<UserProfile>> getParticipantUsers(String participantId, Integer offset, Integer limit) {
+  public ResponseEntity<UserProfiles> getParticipantUsers(String participantId, Integer offset, Integer limit) {
     log.debug("getParticipantUsers.enter; got participantId: {}", participantId);
     List<UserProfile> profiles = partDao.selectUsers(participantId)
         .orElseThrow(() -> new NotFoundException("Participant not found: " + participantId));
     log.debug("getParticipantUsers.exit; returning: {}", profiles.size());
-    return ResponseEntity.ok(profiles);
+    // TODO: set total count
+    return ResponseEntity.ok(new UserProfiles(0, profiles));
   }
 
   /**
@@ -121,20 +144,27 @@ public class ParticipantsService implements ParticipantsApiDelegate {
    *
    * @param offset The number of items to skip before starting to collect the result set. (optional, default to 0)
    * @param limit The number of items to return. (optional, default to 100)
-   * @param orderBy Results will be sorted by this field. (optional)
-   * @param asc Ascending/Descending ordering. (optional, default to true)
    * @return List of registered participants (status code 200)
    *         or May contain hints how to solve the error or indicate what was wrong in the request. (status code 400)
    *         or May contain hints how to solve the error or indicate what went wrong at the server.
    *         Must not outline any information about the internal structure of the server. (status code 500)
    */
   @Override
-  public ResponseEntity<List<Participant>> getParticipants(Integer offset, Integer limit) { //, String orderBy, Boolean asc) {
+  public ResponseEntity<Participants> getParticipants(Integer offset, Integer limit) { //, String orderBy, Boolean asc) {
     // sorting is not supported yet by keycloak admin API
     log.debug("getParticipants.enter; got offset: {}, limit: {}", offset, limit);
-    List<Participant> parts = partDao.search(offset, limit);
-    log.debug("getParticipants.exit; returning: {}", parts.size());
-    return ResponseEntity.ok(parts);
+    List<ParticipantMetaData> partsExt = partDao.search(offset, limit);
+    //Adding actual SD from sd-store for each sd-hash present in keycloak
+    List<Participant> parts =
+        partsExt.stream().map(participant -> {
+          participant.setSelfDescription(
+              selfDescriptionStore.getByHash(participant.getSdHash()).getSelfDescription()
+                  .getContentAsString());
+          return participant;
+        }).collect(
+            Collectors.toList());
+    log.debug("getParticipants.exit; returning size:{}, parts: {}", parts.size(),parts);
+    return ResponseEntity.ok(new Participants(partsExt.size(), parts));
   }
 
   /**
@@ -153,32 +183,41 @@ public class ParticipantsService implements ParticipantsApiDelegate {
   @Transactional
   public ResponseEntity<Participant> updateParticipant(String participantId, String body) {
     log.debug("updateParticipant.enter; got participant: {}", participantId);
-    Participant part = toParticipant(body);
-    part = partDao.update(participantId, part)
+    ParticipantMetaData partExisting = partDao.select(participantId)
         .orElseThrow(() -> new NotFoundException("Participant not found: " + participantId));
-    log.debug("updateParticipant.exit; returning: {}", part);
-    return ResponseEntity.ok(part);
+
+    ContentAccessorDirect contentAccessorDirect = new ContentAccessorDirect(body);
+    VerificationResultParticipant verificationResultParticipant =
+        verificationService.verifyParticipantSelfDescription(contentAccessorDirect);
+
+    log.debug("updateParticipant.existing participant got : {}", partExisting);
+    log.debug("updateParticipant.verificationResultParticipant.got : {}", verificationResultParticipant);
+    if(!SessionUtils.sessionUserHasRole(catalogueAdminRole)  && !partExisting.getId().equals(SessionUtils.getSessionParticipantId())) {
+      throw new ForbiddenException("User has no permissions to update participant : " + participantId);
+    }
+    ParticipantMetaData participantUpdated = getParticipantExtWithValidationAndStore(contentAccessorDirect.getContentAsString());
+
+    ParticipantMetaData participantMetaData = partDao.update(participantId, participantUpdated)
+        .orElseThrow(() -> new NotFoundException("Participant not found: " + participantId));
+    log.debug("updateParticipant.exit; returning: {}", participantMetaData);
+    return ResponseEntity.ok(participantMetaData);
   }
 
   /**
-   * Utility method for converting Participant Self-Description to Participant metadata.
+   * Utility method for converting Participant Self-Description to ParticipantExt metadata.
    *
-   * @param sd Self-Description.
-   * @return Participant metadata.
+   * @param body Self-Description.
+   * @return ParticipantExt ParticipantExt.
    */
-  private Participant toParticipant(String sd) {
-    Map<String, Object> map;
-    try {
-      map = jsonMapper.readValue(sd, MAP_TYPE_REF);
-    } catch (JsonProcessingException ex) {
-      throw new ClientException(ex.getMessage());
-    }
-
-    String id = (String) map.get("id");
-    String name = (String) map.get("holder");
-    Map<String, Object> proof = (Map<String, Object>) map.get("proof");
-    String key = (String) proof.get("verificationMethod");
-    return new Participant(id, name, key, sd);
+  private ParticipantMetaData getParticipantExtWithValidationAndStore(String body) {
+    ContentAccessorDirect contentAccessorDirect = new ContentAccessorDirect(body);
+    VerificationResultParticipant verificationResult = verificationService.verifyParticipantSelfDescription(contentAccessorDirect);
+    log.debug("getParticipantExtWithValidationAndStore.got VerificationResultParticipant: {}",verificationResult);
+    SelfDescriptionMetadata selfDescriptionMetadata = new SelfDescriptionMetadata(contentAccessorDirect, verificationResult);
+    log.debug("getParticipantExtWithValidationAndStore.got SelfDescriptionMetadata: {}",selfDescriptionMetadata);
+    selfDescriptionStore.storeSelfDescription(selfDescriptionMetadata, verificationResult);
+    return new ParticipantMetaData(verificationResult.getId(), verificationResult.getParticipantName(),
+            verificationResult.getParticipantPublicKey(), selfDescriptionMetadata.getSelfDescription().getContentAsString(),
+            selfDescriptionMetadata.getSdHash());
   }
-
 }
